+++ src/new_cfg.h
@@ -69,7 +69,7 @@ void CFG_DeviceGroups_SetRecvFlags(int newSendFlags);
 const char *CFG_DeviceGroups_GetName();
 int CFG_DeviceGroups_GetSendFlags();
 int CFG_DeviceGroups_GetRecvFlags();
-void CFG_SetFlags(unsigned long first4bytes, unsigned long second4bytes);
+void CFG_SetFlags(uint32_t first4bytes, uint32_t second4bytes);
 void CFG_SetFlag(int flag, bool bValue);
 bool CFG_HasFlag(int flag);
 void CFG_SetLoggerFlag(int flag, bool bValue);
